Split generic call validation
