+++ tests/docs_truth/repo_hygiene/mod.rs
@@ -17,6 +17,7 @@ mod resolver_symbol_table;
 mod resolver_validation;
 mod source_truth;
 mod typechecker_behavior_impls;
+mod typechecker_call_validation;
 mod typechecker_imports;
 mod typechecker_program_checking;
 mod typechecker_resolver_validation;
